svn merge -c 1582015 from trunk for HADOOP-10437. Fix the javac warnings in the conf and the util package.
git-svn-id: https://svn.apache.org/repos/asf/hadoop/common/branches/branch-2@1582017 13f79535-47bb-0310-9956-ffa450edef68

+++ hadoop-common-project/hadoop-common/src/main/java/org/apache/hadoop/util/LineReader.java
@@ -245,7 +245,7 @@ public class LineReader implements Closeable {
       }
     } while (newlineLength == 0 && bytesConsumed < maxBytesToConsume);
 
-    if (bytesConsumed > (long)Integer.MAX_VALUE) {
+    if (bytesConsumed > Integer.MAX_VALUE) {
       throw new IOException("Too many bytes before newline: " + bytesConsumed);
     }
     return (int)bytesConsumed;
@@ -343,7 +343,7 @@ public class LineReader implements Closeable {
       }
     } while (delPosn < recordDelimiterBytes.length 
         && bytesConsumed < maxBytesToConsume);
-    if (bytesConsumed > (long) Integer.MAX_VALUE) {
+    if (bytesConsumed > Integer.MAX_VALUE) {
       throw new IOException("Too many bytes before delimiter: " + bytesConsumed);
     }
     return (int) bytesConsumed;
